backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
motor==3.3.2
pymongo==4.6.1
pydantic==2.12.4
numpy
pandas
scipy
folium
requests
geocoder
pydantic-settings==2.1.0
Pillow>=10.0.0
